Extract some calendar config
